feat(desktop): register calibration IPC alongside main process
